Use due date from activity usage, rather than quiz/assignment entity
The assignment due date property probably shouldn't be there, as we've been using entities for due dates as a convention (allows for things like modifying it via actions and whatnot). Rather than adding the same properties.dueDate for quizzes, instead use the `due-date`-classed entity that already exists on both quizzes and assignments.