Rework things to use user activity usages
Looks like this was doing some fairly complicated passing-around of data, centred around us fetching activity usages, rather than _user_ activity usages. We want to fetch _user_ activity usages as activity usages themselves won't have due date overrides on them correctly (among other reasons). Tackling this change meant a fair amount of code could be simplified, so that now we simply fetch user -> fetch my-activities link -> render these results (which are user activity usages) d2l-assessments-list-items.

Another notable change here is that the isOverdue/isCompleted/isEnded stuff is now unambiguous - previously, isOverdue would always be false if the activity was completed by the user, but this is somewhat confusing. Instead, the various state variables are now independent, calculated based on "first principles" (end date/due date/completion entity, etc.) and it's up to the renderer of the userActivityUsageInfo object to figure out what to display correctly (e.g. don't display as overdue if it's complete, even if we are past the due date).

This also required a fairly large rework of the tests (by which I mean, removal of tests). This is mostly just because the new code is simpler, since it's skipping a bunch of no-longer-needed checks and stuff, but more tests will be incoming to fill in the gaps that this left.