Skip unsupported user activity usages
This came up in testing the discussions activities API - due to `_getActivityRequest` explicitly looking for an assignment or quiz, we end up trying to fetching undefined URLs, and then bad things happen. Instead, take a similar approach that is being done with the overdue activities already - filter the user activity usages down to only the types we know how to handle up front, and avoid said bad things!